+++ src/KingmakerGunslinger/Blueprints/GunslingerClassBlueprints.cs
@@ -30,7 +30,8 @@ namespace KingmakerGunslinger.Blueprints
             BleedingWoundBlueprintSet bleedingWound,
             ExpertLoadingBlueprintSet expertLoading,
             LightningReloadBlueprintSet lightningReload,
-            EvasiveBlueprintSet evasive, MenacingShotBlueprintSet menacingShot)
+            EvasiveBlueprintSet evasive, MenacingShotBlueprintSet menacingShot,
+            SlingersLuckBlueprintSet slingersLuck)
         {
             CharacterClass = characterClass ?? throw new ArgumentNullException("characterClass");
             Progression = progression ?? throw new ArgumentNullException("progression");
@@ -62,6 +63,8 @@ namespace KingmakerGunslinger.Blueprints
             Evasive = evasive ?? throw new ArgumentNullException("evasive");
             MenacingShot = menacingShot ??
                 throw new ArgumentNullException("menacingShot");
+            SlingersLuck = slingersLuck ??
+                throw new ArgumentNullException("slingersLuck");
         }
         internal BlueprintCharacterClass CharacterClass { get; private set; }
         internal BlueprintProgression Progression { get; private set; }
@@ -85,7 +88,8 @@ namespace KingmakerGunslinger.Blueprints
         internal LightningReloadBlueprintSet LightningReload { get; private set; }
         internal EvasiveBlueprintSet Evasive { get; private set; }
         internal MenacingShotBlueprintSet MenacingShot { get; private set; }
-        internal int Count { get { return 4 + Grit.Count + Deadeye.Count + Dodge.Count + QuickClear.Count + Nimble.Count + PistolWhip.Count + UtilityShot.Count + GunTraining.Count + DeadShot.Count + StartlingShot.Count + TargetingHead.Count + TargetingTorso.Count + TargetingLegs.Count + BleedingWound.Count + ExpertLoading.Count + LightningReload.Count + Evasive.Count + MenacingShot.Count; } }
+        internal SlingersLuckBlueprintSet SlingersLuck { get; private set; }
+        internal int Count { get { return 4 + Grit.Count + Deadeye.Count + Dodge.Count + QuickClear.Count + Nimble.Count + PistolWhip.Count + UtilityShot.Count + GunTraining.Count + DeadShot.Count + StartlingShot.Count + TargetingHead.Count + TargetingTorso.Count + TargetingLegs.Count + BleedingWound.Count + ExpertLoading.Count + LightningReload.Count + Evasive.Count + MenacingShot.Count + SlingersLuck.Count; } }
     }
 
     internal sealed class GunslingerClassCatalogPublication
@@ -195,6 +199,9 @@ namespace KingmakerGunslinger.Blueprints
                 registry, grit.Resource, characterClass);
             MenacingShotBlueprintSet menacingShot =
                 MenacingShotBlueprints.Register(library, registry);
+            SlingersLuckBlueprintSet slingersLuck =
+                SlingersLuckBlueprints.Register(registry, grit.Resource,
+                    characterClass);
             BlueprintProgression progression = registry.Register<BlueprintProgression>(
                 ProgressionSymbol, () => CreateProgression());
 
@@ -219,6 +226,7 @@ namespace KingmakerGunslinger.Blueprints
             progression.LevelEntries[10].Features.Add(lightningReload.Feature);
             progression.LevelEntries[14].Features.Add(evasive.Feature);
             progression.LevelEntries[14].Features.Add(menacingShot.Feature);
+            progression.LevelEntries[14].Features.Add(slingersLuck.Feature);
             Validate(characterClass, progression, proficiencies, fullBab, goodSave,
                 poorSave, startingPistol, blackPowder, leadBall,
                 simple, martial, lightArmor, firearmProficiency);
@@ -226,7 +234,8 @@ namespace KingmakerGunslinger.Blueprints
                 proficiencies, grit, deadeye, dodge, quickClear, nimble, initiative,
                 pistolWhip, utilityShot, gunTraining, deadShot, startlingShot,
                 targetingHead, targetingTorso, targetingLegs, bleedingWound,
-                expertLoading, lightningReload, evasive, menacingShot);
+                expertLoading, lightningReload, evasive, menacingShot,
+                slingersLuck);
         }
 
         internal static GunslingerClassCatalogPublication Publish(